requirements.txt
# Clinical Decision Support System — Python Dependencies
# Python 3.11+

# ── Google ADK ───────────────────────────────────────────────────────────────
google-adk>=0.1.0

# ── Google Cloud ─────────────────────────────────────────────────────────────
google-cloud-healthcare>=1.4.0
google-cloud-pubsub>=2.21.0
google-cloud-dlp>=3.15.0
google-cloud-kms>=2.21.0
google-cloud-firestore>=2.15.0
google-cloud-bigquery>=3.17.0
google-cloud-storage>=2.16.0
google-cloud-logging>=3.9.0
google-cloud-aiplatform>=1.51.0

# ── Vertex AI ────────────────────────────────────────────────────────────────
vertexai>=1.51.0

# ── Vertex AI Search ─────────────────────────────────────────────────────────
google-cloud-discoveryengine>=0.11.0

# ── Auth ─────────────────────────────────────────────────────────────────────
google-auth>=2.29.0
google-auth-httplib2>=0.2.0

# ── Data validation ──────────────────────────────────────────────────────────
pydantic>=2.6.0

# ── HTTP ─────────────────────────────────────────────────────────────────────
requests>=2.31.0
httpx>=0.27.0

# ── Environment ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
